+++ src/com/lawknow/userWrite/UserWriteOk.java
@@ -24,7 +24,7 @@ public class UserWriteOk implements Action {
 		HttpSession session = req.getSession();
 		ActionInfo actionInfo = new ActionInfo();
 		
-		String uploadPath ="D:\\web_1900_kgw\\jsp\\workspace\\kovengerss\\WebContent\\upload"; //여기다가 첨부된 파일을 업로드 하겠다
+		String uploadPath ="C:\\web_1900_kgw\\jsp\\kovengerss\\WebContent\\upload"; //여기다가 첨부된 파일을 업로드 하겠다
 		int fileSize =1024*1024*5; //5M
 		
 		UserVO userVO = new UserVO();
